nonprofit_coach/requirements.txt
Flask==3.0.0
anthropic==0.39.0
python-dotenv==1.0.0
